+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 [Full Changelog](https://github.com/uktrade/great-domestic-ui/compare/2019.07.16...2019.07.23)
 
 ### Implemented enhancements
+- TT-1678 - Rename EU exit to Brexit
 - XOT-958 - bugfixes for XOT-922 (css changes, added cms_breadcrumbs, fixed template footer markup and hide block if fields not populated)
 - XOT-914 - Add automated tests for search sort order
 - XOT-906 - Update wording on search feedback form